+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Go to the `v1` branch to see the changelog of Lume 1.
 ## [3.0.10] - Unreleased
 ### Added
 - New env variable `LUME_PROXIED` to use a location with a different hostname and port to the server's hostname and port.
+- New file `serve.ts` to run LumeCMS in a VPS.
 - Ignore `deno.json` and `deno.jsonc` files by the watcher.
 
 ### Fixed